# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Documentation for Cric Heros",
        "title": "Cric Heros API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "paths": {
...
        "/addToScoreCard": {
            "post": {
                "description": "stores players info in scorecard",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Scorecard"
                ],
                "parameters": [
                    {
                        "description": "ScoreCard details",
                        "name": "details",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CardData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
        "/createPlayer": {
            "post": {
                "description": "Creates a new Player",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Player"
                ],
                "parameters": [
                    {
                        "description": "Create Player",
                        "name": "player",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Player"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
        "/retirePlayer": {
            "delete": {
                "description": "Shows the list of all the player",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Player"
                ],
                "parameters": [
                    {
                        "description": "Id of the player",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
        "/showPlayer": {
            "get": {
                "description": "Shows the list of all the player",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Player"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/showPlayerID": {
            "get": {
                "description": "Shows the list of all the player",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Player"
                ],
                "parameters": [
                    {
                        "description": "ID of the player",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/showScoreCard": {
            "post": {
                "description": "Shows the score card for the current matcha",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Scorecard"
                ],
                "parameters": [
                    {
                        "description": "Id of the match whose scoredcard is to be viewed",
                        "name": "match_id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "models.CardData": {
            "type": "object",
            "required": [
                "ballType",
                "baller",
                "batsmen",
                "matchId"
            ],
            "properties": {
                "ballType": {
                    "type": "string"
                },
                "baller": {
                    "type": "string"
                },
                "batsmen": {
                    "type": "string"
                },
                "matchId": {
                    "type": "string"
                },
                "prevRuns": {
                    "type": "integer"
                },
                "runs": {
                    "type": "integer"
                }
            }
        },
...
        "models.Player": {
            "type": "object",
            "required": [
                "country",
                "jerseyNo",
                "phoneNo",
                "playerAge",
                "playerName"
            ],
            "properties": {
                "country": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "jerseyNo": {
                    "type": "integer"
                },
                "phoneNo": {
                    "type": "string"
                },
                "playerAge": {
                    "type": "integer"
                },
                "playerId": {
                    "type": "string"
                },
                "playerName": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "status": {
                    "type": "string"
                }
            }
        },
...
    }
}